Build MP4 container with MP4Box (#30)
Ship Broadcast via DASH #29

Side lab experiments done in
 - https://github.com/charneykaye/encode-fmp4-demo

referenced in the question I posted on Stack Overflow
 - https://stackoverflow.com/questions/69625970/java-mp4parser-to-create-a-single-m4s-fragment

+++ lib/src/main/java/io/xj/lib/mixer/AudioSampleFormat.java
@@ -423,7 +423,7 @@ public enum AudioSampleFormat {
    @return value
    */
   private static double fromBytesF32LSB(byte[] sample) {
-    return (double) ByteBuffer.wrap(sample)
+    return ByteBuffer.wrap(sample)
       .order(ByteOrder.LITTLE_ENDIAN)
       .getFloat();
   }
@@ -435,7 +435,7 @@ public enum AudioSampleFormat {
    @return value
    */
   private static double fromBytesF32MSB(byte[] sample) {
-    return (double) ByteBuffer.wrap(sample)
+    return ByteBuffer.wrap(sample)
       .order(ByteOrder.BIG_ENDIAN)
       .getFloat();
   }
